+++ packages/save-and-share-bar/src/tracking/with-track-events.js
@@ -1,0 +1,52 @@
+import { withTrackEvents } from "@times-components/tracking";
+
+export default Component =>
+  withTrackEvents(Component, {
+    analyticsEvents: [
+      {
+        actionName: "Pressed",
+        eventName: "onSaveToMyArticles",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline
+        }),
+        trackingName: "ArticleSaveToMyArticles"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onShareOnEmail",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline
+        }),
+        trackingName: "ArticleShareOnEmail"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onCopyLink",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline
+        }),
+        trackingName: "ArticleCopyLink"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onShareOnFB",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline
+        }),
+        trackingName: "ArticleShareOnFacebook"
+      },
+      {
+        actionName: "Pressed",
+        eventName: "onShareOnTwitter",
+        getAttrs: ({ articleId, articleHeadline }) => ({
+          articleId,
+          articleHeadline
+        }),
+        trackingName: "ArticleShareOnTwitter"
+      }
+    ]
+  });