+++ packages/article/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.1.1](https://github.com/newsuk/times-components/compare/@times-components/article@7.1.0...@times-components/article@7.1.1) (2019-05-13)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 # [7.1.0](https://github.com/newsuk/times-components/compare/@times-components/article@7.0.16...@times-components/article@7.1.0) (2019-05-10)